+++ packages/inventor-plugin-plugin/templates/default/src/index.ts
@@ -4,6 +4,6 @@
  */
  import { Plugin } from '@inventorjs/cli-core'
 
- export default class extends Plugin {
+ export default class <%- pluginClassName %> extends Plugin {
    description = '<%- description %>'
  }