+++ packages/inventor-plugin-app-react-webpack/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @inventorjs/plugin-app-react-webpack
 
+## 5.0.13
+
+### Patch Changes
+
+- support argument
+- Updated dependencies
+  - @inventorjs/cli-core@5.0.13
+
 ## 5.0.12
 
 ### Patch Changes